requirements.txt
Babel==2.6.0
boto3==1.9.86
botocore==1.12.86
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
confluent-kafka==0.11.6
docutils==0.14
elasticsearch==6.3.1
google-api-core==1.7.0
google-auth==1.6.2
google-cloud==0.34.0
google-cloud-core==0.29.1
google-cloud-storage==1.13.2
google-resumable-media==0.3.2
googleapis-common-protos==1.5.5
idna==2.8
iso8601==0.1.12
jmespath==0.9.3
kafka-python==1.4.4
kazoo==2.5.0
msgpack==0.6.0
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
python-dateutil==2.7.5
pytimeparse==1.1.8
pytz==2018.7
requests==2.21.0
rsa==4.0
s3transfer==0.1.13
six==1.12.0
tabulate==0.8.3
ujson==1.35
urllib3==1.24.1
zulu==1.1.0
